infrastructure/tests/requirements.txt
pytest
azure-cli